+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/CounterTest.kt
@@ -33,18 +33,9 @@ class CounterTest {
 	@Test fun counterWithAnsi() = runTest {
 		runMosaicTest(withAnsi = true) {
 			setCounter()
-			assertThat(awaitRenderSnapshot()).isEqualTo(
-				"""
-				|${ansiBeginSynchronizedUpdate}The count is: 0
-				|$ansiEndSynchronizedUpdate
-				""".trimMargin().replaceLineEndingsWithCRLF(),
-			)
-			for (i in 1..20) {
+			for (i in 0..20) {
 				assertThat(awaitRenderSnapshot()).isEqualTo(
-					"""
-					|${ansiBeginSynchronizedUpdate}${cursorUp}The count is: ${i}$clearLine
-					|$ansiEndSynchronizedUpdate
-					""".trimMargin().replaceLineEndingsWithCRLF(),
+					"${ansiMoveCursorToFirstColumn}The count is: ${i}$ansiClearLineAfterCursor$ansiClearAllAfterCursor".wrapWithAnsiSynchronizedUpdate(),
 				)
 			}
 		}