+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/MosaicTest.kt
@@ -45,10 +45,9 @@ class MosaicTest {
 		}
 		assertThat(actual).isEqualTo(
 			"""
-			|One $s
-			|Two $s
-			|Three
-			|
+			|${ansiMoveCursorToFirstColumn}One  $ansiClearLineAfterCursor
+			|Two  $ansiClearLineAfterCursor
+			|Three$ansiClearLineAfterCursor$ansiClearAllAfterCursor
 			""".trimMargin().wrapWithAnsiSynchronizedUpdate().replaceLineEndingsWithCRLF(),
 		)
 	}
@@ -75,10 +74,9 @@ class MosaicTest {
 		}
 		assertThat(actual).isEqualTo(
 			"""
-			|One $s
-			|Two $s
-			|Three
-			|
+			|${ansiMoveCursorToFirstColumn}One  $ansiClearLineAfterCursor
+			|Two  $ansiClearLineAfterCursor
+			|Three$ansiClearLineAfterCursor$ansiClearAllAfterCursor
 			""".trimMargin().wrapWithAnsiSynchronizedUpdate().replaceLineEndingsWithCRLF(),
 		)
 	}
@@ -105,10 +103,9 @@ class MosaicTest {
 		}
 		assertThat(actual).isEqualTo(
 			"""
-			|One $s
-			|Two $s
-			|Three
-			|
+			|${ansiMoveCursorToFirstColumn}One  $ansiClearLineAfterCursor
+			|Two  $ansiClearLineAfterCursor
+			|Three$ansiClearLineAfterCursor$ansiClearAllAfterCursor
 			""".trimMargin().wrapWithAnsiSynchronizedUpdate().replaceLineEndingsWithCRLF(),
 		)
 	}
@@ -147,10 +144,9 @@ class MosaicTest {
 		}
 		assertThat(actual).isEqualTo(
 			"""
-			|One $s
-			|Two $s
-			|Three
-			|
+			|${ansiMoveCursorToFirstColumn}One  $ansiClearLineAfterCursor
+			|Two  $ansiClearLineAfterCursor
+			|Three$ansiClearLineAfterCursor$ansiClearAllAfterCursor
 			""".trimMargin().wrapWithAnsiSynchronizedUpdate().replaceLineEndingsWithCRLF(),
 		)
 	}
@@ -167,10 +163,9 @@ class MosaicTest {
 		}
 		assertThat(actual).isEqualTo(
 			"""
-			|One $s
-			|Two $s
-			|Three
-			|
+			|${ansiMoveCursorToFirstColumn}One  $ansiClearLineAfterCursor
+			|Two  $ansiClearLineAfterCursor
+			|Three$ansiClearLineAfterCursor$ansiClearAllAfterCursor
 			""".trimMargin().wrapWithAnsiSynchronizedUpdate().replaceLineEndingsWithCRLF(),
 		)
 	}
@@ -191,10 +186,9 @@ class MosaicTest {
 		actuals.forEach { actual ->
 			assertThat(actual).isEqualTo(
 				"""
-				|One $s
-				|Two $s
-				|Three
-				|
+				|${ansiMoveCursorToFirstColumn}One  $ansiClearLineAfterCursor
+				|Two  $ansiClearLineAfterCursor
+				|Three$ansiClearLineAfterCursor$ansiClearAllAfterCursor
 				""".trimMargin().wrapWithAnsiSynchronizedUpdate().replaceLineEndingsWithCRLF(),
 			)
 		}